Build a lock-less QtSerialPort
Qt tries to lock the serial port for opening, but the paths it's trying are inaccessible,
and don't make sense anyway since the locks wouldn't be visible outside the snap.

The only way to avoid this error seems to be building a version of Qt without locking.

Ref. https://forum.snapcraft.io/t/access-to-var-lock-or-alternatives/21269